recombine
